split the code in remote_decorator.py; add proxy_wrapper test; block get attribute in future mode
