+++ parl/remote/client.py
@@ -248,7 +248,7 @@ class Client(object):
 
     def _check_and_monitor_job(self, job_heartbeat_address,
                                ping_heartbeat_address, max_memory,
-                               remote_wrapper_object):
+                               proxy_wrapper_nowait_object):
         """ Sometimes the client may receive a job that is dead, thus 
         we have to check if this job is still alive before adding it to the `actor_num`.
         """
@@ -276,14 +276,15 @@ class Client(object):
         # a thread for sending heartbeat signals to job
         thread = threading.Thread(
             target=self._create_job_monitor,
-            args=(job_heartbeat_socket, remote_wrapper_object))
+            args=(job_heartbeat_socket, proxy_wrapper_nowait_object))
         thread.setDaemon(True)
         thread.start()
         return True
 
-    def _create_job_monitor(self, job_heartbeat_socket, remote_wrapper_object):
+    def _create_job_monitor(self, job_heartbeat_socket,
+                            proxy_wrapper_nowait_object):
         """Send heartbeat signals to check target's status"""
-        copy_remote_wrapper_object = remote_wrapper_object
+        copy_proxy_wrapper_nowait_object = proxy_wrapper_nowait_object
 
         # @TODO(zenghsh3): hardcoded
         actor_deleted_refcount = 7
@@ -292,8 +293,9 @@ class Client(object):
 
         job_is_alive = True
         while job_is_alive and self.client_is_alive:
-            if copy_remote_wrapper_object is not None:
-                cur_refcount = sys.getrefcount(copy_remote_wrapper_object)
+            if copy_proxy_wrapper_nowait_object is not None:
+                cur_refcount = sys.getrefcount(
+                    copy_proxy_wrapper_nowait_object)
                 if cur_refcount == actor_deleted_refcount:
                     # actor is deleted or out of scope, we will terminate the heartbeat of the job
                     # and release the cpu resource.
@@ -330,7 +332,7 @@ class Client(object):
 
         job_heartbeat_socket.close(0)
 
-    def submit_job(self, max_memory, remote_wrapper_object):
+    def submit_job(self, max_memory, proxy_wrapper_nowait_object):
         """Send a job to the Master node.
 
         When a `@parl.remote_class` object is created, the global client
@@ -341,9 +343,9 @@ class Client(object):
             max_memory (float): Maximum memory (MB) can be used by each remote
                                 instance, the unit is in MB and default value is
                                 none(unlimited).
-            remote_wrapper_object (object): instance of `RemoteWrapper`, use the 
-                                reference count of the object to detack whether 
-                                the remote_class has been deleted or out of scope.
+            proxy_wrapper_nowait_object (object): instance of actor class which is decorated by @remote_class(wait=False),
+                                use the reference count of the object to detect whether 
+                                the object has been deleted or out of scope.
 
         Returns:
             job_address(str): IP address of the job. None if there is no available CPU in the cluster.
@@ -370,7 +372,7 @@ class Client(object):
 
                     check_result = self._check_and_monitor_job(
                         job_heartbeat_address, ping_heartbeat_address,
-                        max_memory, remote_wrapper_object)
+                        max_memory, proxy_wrapper_nowait_object)
                     if check_result:
                         self.lock.acquire()
                         self.actor_num += 1
